+++ src/LurchServerCommand.ml
@@ -114,8 +114,8 @@ let do_monitor_run isolation_run cgroup pid stdout_r stderr_r timeout run_id =
         log.info "Consumed %a+%a CPU and %a+%a RAM"
           (Option.print Float.print) cpu_usr
           (Option.print Float.print) cpu_sys
-          (Option.print Int.print) mem_usr
-          (Option.print Int.print) mem_sys ;
+          (Option.print Int64.print) mem_usr
+          (Option.print Int64.print) mem_sys ;
         Db.Run.stop run_id exit_code
                     ?cpu_usr ?cpu_sys ?mem_usr ?mem_sys ;
         (* Also deletes the cgroup: *)
